Refactor PR comments
