+++ src/orders/components/OrderTransactionEvents/EventTime.tsx
@@ -1,0 +1,18 @@
+import useLocale from "@saleor/hooks/useLocale";
+import React from "react";
+
+const EventTime: React.FC<{ date: string }> = ({ date }) => {
+  const { locale } = useLocale();
+  const intl = new Intl.DateTimeFormat(locale, {
+    timeZoneName: "short",
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+
+  return <time dateTime={date}>{intl.format(new Date(date))}</time>;
+};
+
+export default EventTime;
